Showing off, indexed 2d arrays
It can be challenging to work with 2d arrays if they are used to model
objects.

Using typed dimensions helps, but you can also add explicit indexes to
the data, which makes it easier to look read the raw DZN data.

The other alternative is to use a list of record, where each record has
an attribute (lenght, width, height) for each product.

So something like:

```
product_size = [
  P(1): ( length: 4, width: 4, height: 4 ),
  P(2): ( length: 8, width: 4, height: 4 ),
];
```